Restructured web_api_testing
