+++ src/hackingBuddyGPT/usecases/web_api_testing/documentation/parsing/openapi_converter.py
