+++ src/hackingBuddyGPT/usecases/web_api_testing/simple_openapi_documentation.py
@@ -8,7 +8,7 @@ from hackingBuddyGPT.capabilities.record_note import RecordNote
 from hackingBuddyGPT.usecases.agents import Agent
 from hackingBuddyGPT.usecases.web_api_testing.prompt_generation.information.prompt_information import PromptContext
 from hackingBuddyGPT.usecases.web_api_testing.utils.custom_datatypes import Prompt, Context
-from hackingBuddyGPT.usecases.web_api_testing.utils.documentation.openapi_specification_handler import OpenAPISpecificationHandler
+from hackingBuddyGPT.usecases.web_api_testing.documentation.openapi_specification_handler import OpenAPISpecificationHandler
 from hackingBuddyGPT.usecases.web_api_testing.utils.llm_handler import LLMHandler
 from hackingBuddyGPT.usecases.web_api_testing.prompt_generation.prompt_engineer import PromptStrategy, PromptEngineer
 from hackingBuddyGPT.usecases.web_api_testing.response_processing.response_handler import ResponseHandler
