+++ src/hackingBuddyGPT/usecases/web_api_testing/simple_web_api_testing.py
@@ -11,8 +11,8 @@ from hackingBuddyGPT.capabilities.record_note import RecordNote
 from hackingBuddyGPT.usecases.agents import Agent
 from hackingBuddyGPT.usecases.web_api_testing.prompt_generation.information.prompt_information import PromptContext
 from hackingBuddyGPT.usecases.web_api_testing.utils.custom_datatypes import Prompt, Context
-from hackingBuddyGPT.usecases.web_api_testing.utils.documentation.parsing import OpenAPISpecificationParser
-from hackingBuddyGPT.usecases.web_api_testing.utils.documentation.report_handler import ReportHandler
+from hackingBuddyGPT.usecases.web_api_testing.documentation.parsing import OpenAPISpecificationParser
+from hackingBuddyGPT.usecases.web_api_testing.documentation.report_handler import ReportHandler
 from hackingBuddyGPT.usecases.web_api_testing.utils.llm_handler import LLMHandler
 from hackingBuddyGPT.usecases.web_api_testing.prompt_generation.prompt_engineer import PromptEngineer, PromptStrategy
 from hackingBuddyGPT.usecases.web_api_testing.response_processing.response_handler import ResponseHandler
