+++ src/hackingBuddyGPT/usecases/web_api_testing/utils/__init__.py
@@ -1,1 +1,2 @@
 from .llm_handler import LLMHandler
+from .custom_datatypes import Prompt,  Context
