+++ tests/test_openapi_converter.py
@@ -2,7 +2,7 @@ import unittest
 from unittest.mock import patch, mock_open
 import os
 
-from hackingBuddyGPT.usecases.web_api_testing.utils.documentation.parsing.openapi_converter import OpenAPISpecificationConverter
+from hackingBuddyGPT.usecases.web_api_testing.documentation.parsing.openapi_converter import OpenAPISpecificationConverter
 
 
 class TestOpenAPISpecificationConverter(unittest.TestCase):
